folder rename and normalize naming processor.py
