+++ .env.example
@@ -1,6 +1,6 @@
 MESSAGING_TYPE=RabbitMqConnection
 PROCESSOR_TYPE=LoggingServiceProcessor
-PROCESSOR_MODULE=services.logging_service_processor
+PROCESSOR_MODULE=services.processor
 
 RABBITMQ_QUEUE=test_queue
 RABBITMQ_VIRTUAL_HOST=/
